# PERF: improve 2D array access / transpose() for masked dtypes