xref https://github.com/pandas-dev/pandas/issues/52016

For `transpose()` or operations that do a transpose under the hood (eg typically operations with axis=1), this is generally expensive if you have extension dtypes. This explores the possibilities to add a fast path for the specific case of a DataFrame with all-masked dtypes.

Using the example dataframe from https://github.com/pandas-dev/pandas/issues/52016, this gives a speed-up 

```python
shape = 250_000, 100
mask = pd.DataFrame(np.random.randint(0, 1, size=shape))

np_mask = mask.astype(bool)
pd_mask = mask.astype(pd.BooleanDtype())

%timeit pd_mask.transpose()
# 9.13 s ± 347 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)   <-- main
# 689 ms ± 28.5 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  <-- PR
```


Note there are some various different optimizations added together (that can be splitted, and still needs to be cleaned-up in general). 
The main speed-up comes from calling `BooleanArray(..)` instead of `BooleanArray._from_sequence(..)` (or in general with any other masked array class), because this generic `_from_sequence` has a lot of overhead (that adds up when calling it 250_000 times as in this example). To be able to use the main `__init__`, I added a version of `mgr.as_array()` to return separate data and mask arrays. 

Further, after the arrays are constructed, the BlockManager construction can be optimized by passing `verify_integrity=False` and by knowing that we have all 1D EAs (so no need to do block dtype grouping etc). 
Another tiny optimization is in the BooleanArray(..) to avoid creating a new dtype instance every time (this can easily be split of, we might want to do that generally for all primitive dtypes, and maybe rather in the init of the dtype)