@jorisvandenbossche 2023-03-20
Yes, I also mentioned that in https://github.com/pandas-dev/pandas/issues/52016#issuecomment-1476158450, but I think we can certainly try to avoid transpose altogether, but improving transpose itself as well might also be worth it? (edit: plus for methods where the transpose cannot easily be avoided, as Richard just commented)  
(the main new function here that I added, `as_array_masked`, could also be used for both, see my inline comment above at https://github.com/pandas-dev/pandas/pull/52083#discussion_r1142041025)

> Can any of this be extended eventually to arrow-backed cases?

Yes, in general the arrow dtypes will suffer the same problems for row based operations (as long as arrow doesn't add some specific functionality for that). I think it could be quite easy to expand the approach here to work for arrow dtypes as well (converting their bitmask to a bytemask)